+++ app/Http/Controllers/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Role;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UsersController extends Controller
 {
@@ -13,10 +14,14 @@ class UsersController extends Controller
      */
     public function index()
     {
+        if (!user_can('users_access')) {
+            return redirect()->route('dashboard');
+        }
+
         // $users = User::OrderByDesc('created_at')->first();
         $users = User::OrderByDesc('created_at')->get();
 
-        return view('users.index')->with([
+        return view('admin.users.index')->with([
             'users' => $users
         ]);
     }
@@ -26,9 +31,13 @@ class UsersController extends Controller
      */
     public function create()
     {
-        $roles = Role::orderByDesc('created_at')->get();
+        if (!user_can('users_access')) {
+            return redirect()->route('dashboard');
+        }
 
-        return view('users.form')->with([
+        $roles = Role::where('id', '!=', 1)->orderByDesc('created_at')->get();
+
+        return view('admin.users.form')->with([
             'roles' => $roles
         ]);
     }
@@ -38,12 +47,22 @@ class UsersController extends Controller
      */
     public function store(Request $request)
     {
+        if (!user_can('users_access')) {
+            return redirect()->route('dashboard');
+        }
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|unique:users,email',
             'password' => 'required|string',
             'role_id' => 'required|exists:roles,id'
         ]);
+
+        if ($validated['role_id'] == 1) {
+            return back()->with([
+                'error' => 'Failed to create user. Cannot add Super Admin Roles.'
+            ]);
+        }
 
         $validated['password'] = bcrypt($validated['password']);
 
@@ -63,9 +82,14 @@ class UsersController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show()
     {
-        //
+      
+        $user = Auth::user();
+
+        return view('admin.users.show')->with([
+            'user' => $user
+        ]);
     }
 
     /**
@@ -73,10 +97,22 @@ class UsersController extends Controller
      */
     public function edit(string $id)
     {
+        if (!user_can('users_access')) {
+            return redirect()->route('dashboard');
+        }
+
+
+
         $result = User::findOrFail($id);
+
+
+        if ($result->role->id == 1) {
+            return redirect()->route('users.index');
+        }
+
         $roles = Role::orderByDesc('created_at')->get();
 
-        return view('users.form')->with([
+        return view('admin.users.form')->with([
             'result' => $result,
             'roles' => $roles
         ]);
@@ -87,12 +123,21 @@ class UsersController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!user_can('users_access')) {
+            return redirect()->route('dashboard');
+        }
 
         $user = User::findOrFail($id);
 
+        if ($user->role->id == 1) {
+            return back()->with([
+                'error' => 'Failed to update user. Cannot update Super Admin.'
+            ]);
+        }
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
-            'email' => 'required|unique:users,email,'.$user->id,
+            'email' => 'required|unique:users,email,' . $user->id,
             'role_id' => 'required|exists:roles,id'
         ]);
 
@@ -119,7 +164,17 @@ class UsersController extends Controller
      */
     public function destroy(string $id)
     {
+        if (!user_can('users_access')) {
+            return redirect()->route('dashboard');
+        }
+
         $user = User::findOrFail($id);
+
+        if ($user->role->id == 1) {
+            return back()->with([
+                'error' => 'Failed to deleting user. Cannot delete Super Admin.'
+            ]);
+        }
 
         $user->delete($id);
 
@@ -128,5 +183,50 @@ class UsersController extends Controller
                 'success' => 'User Successfully Deleted.'
             ]);
         }
+    }
+
+
+
+    // profile
+
+    public function editProfile()
+    {
+       
+
+        $user = Auth::user();
+        $roles = Role::orderByDesc('created_at')->get();
+
+
+        return view('admin.users.show-form')->with([
+            'user' => $user,
+            'roles' => $roles
+        ]);
+    }
+
+    public function updateProfile(Request $request)
+    {
+       
+
+        $userId = Auth::user()->id;
+        $user = User::findOrFail($userId);
+
+        $validated = $request->validate([
+            'name' => 'sometimes|string|max:255',
+            'email' => 'sometimes|string|email|unique:users,email,' . Auth::id(),
+            'new_password' => 'nullable|string|min:6',
+            'role_id' => 'sometimes|exists:roles,id',
+        ]);
+
+        $status = $user->update($validated);
+
+        if (!$status) {
+            return back()->with([
+                'error' => "Failed to updating Profile."
+            ]);
+        }
+
+        return redirect()->route('user.show')->with([
+            'success' => 'Profile Successfully updated.'
+        ]);
     }
 }
